+++ lib/stellar/accounts.ex
@@ -11,4 +11,12 @@ defmodule Stellar.Accounts do
   def get(accountId) do
     Base.get("/accounts/#{accountId}")
   end
+
+  @doc """
+  Gets a single data associated with the given account
+  """
+  @spec get_data(binary, binary) :: {Stellar.status, map}
+  def get_data(accountId, key) do
+    Base.get("/accounts/#{accountId}/data/#{key}")
+  end
 end